+++ src/main/java/co/edu/unbosque/repository/PresupuestoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.unbosque.model.Presupuesto;
 
-public interface PresupuestoRepository extends JpaRepository<Presupuesto, Integer>{
+public interface PresupuestoRepository extends JpaRepository<Presupuesto, Long>{
 
 }
