+++ src/main/java/co/edu/unbosque/service/api/CategoriaServiceAPI.java
@@ -3,6 +3,6 @@ package co.edu.unbosque.service.api;
 import co.edu.unbosque.model.Categoria;
 import co.edu.unbosque.utils.GenericServiceAPI;
 
-public interface CategoriaServiceAPI extends GenericServiceAPI<Categoria, Integer>{
+public interface CategoriaServiceAPI extends GenericServiceAPI<Categoria, Long> {
 
 }
