+++ src/main/java/co/edu/unbosque/service/api/ConceptoServiceAPI.java
@@ -3,6 +3,6 @@ package co.edu.unbosque.service.api;
 import co.edu.unbosque.model.Concepto;
 import co.edu.unbosque.utils.GenericServiceAPI;
 
-public interface ConceptoServiceAPI extends GenericServiceAPI<Concepto, Integer> {
+public interface ConceptoServiceAPI extends GenericServiceAPI<Concepto, Long> {
 
 }
