+++ src/main/java/co/edu/unbosque/service/api/PresupuestoServiceAPI.java
@@ -3,6 +3,6 @@ package co.edu.unbosque.service.api;
 import co.edu.unbosque.model.Presupuesto;
 import co.edu.unbosque.utils.GenericServiceAPI;
 
-public interface PresupuestoServiceAPI extends GenericServiceAPI<Presupuesto, Integer> {
+public interface PresupuestoServiceAPI extends GenericServiceAPI<Presupuesto, Long> {
 
 }
