+++ src/main/java/co/edu/unbosque/service/imp/PresupuestoServiceImpl.java
@@ -10,13 +10,13 @@ import co.edu.unbosque.service.api.PresupuestoServiceAPI;
 import co.edu.unbosque.utils.GenericServiceImpl;
 
 @Service
-public class PresupuestoServiceImpl extends GenericServiceImpl<Presupuesto, Integer>  implements PresupuestoServiceAPI{
+public class PresupuestoServiceImpl extends GenericServiceImpl<Presupuesto, Long>  implements PresupuestoServiceAPI{
 
 	@Autowired
 	private PresupuestoRepository presupuestoDaoAPI;
 	
 	@Override
-	public CrudRepository<Presupuesto, Integer> getDao() {
+	public CrudRepository<Presupuesto, Long> getDao() {
 		return presupuestoDaoAPI;
 	}
